Refactored HUD code, HUD elements can now be registered from other features.
