+++ spt.vcxproj.filters
@@ -208,6 +208,9 @@
     <ClCompile Include="spt\cvars.cpp">
       <Filter>spt</Filter>
     </ClCompile>
+    <ClCompile Include="spt\features\hops_hud.cpp">
+      <Filter>spt\features</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\public\tier0\basetypes.h">
@@ -336,9 +339,6 @@
     <ClInclude Include="spt\feature.hpp">
       <Filter>spt</Filter>
     </ClInclude>
-    <ClInclude Include="spt\features\isg.hpp">
-      <Filter>spt\features</Filter>
-    </ClInclude>
     <ClInclude Include="spt\features\shadow.hpp">
       <Filter>spt\features</Filter>
     </ClInclude>
@@ -463,6 +463,9 @@
       <Filter>spt\vgui</Filter>
     </ClInclude>
     <ClInclude Include="spt\features\tas.hpp">
+      <Filter>spt\features</Filter>
+    </ClInclude>
+    <ClInclude Include="spt\features\hud.hpp">
       <Filter>spt\features</Filter>
     </ClInclude>
   </ItemGroup>
