+++ spt/features/isg.cpp
@@ -1,4 +1,5 @@
 #include "stdafx.h"
+#include "hud.hpp"
 #include "..\feature.hpp"
 #include "..\utils\game_detection.hpp"
 
@@ -7,7 +8,9 @@
 
 #include "convar.hpp"
 
-#if defined(SSDK2007) || defined(SSDK2013)
+#if defined(SSDK2007)
+
+ConVar y_spt_hud_isg("y_spt_hud_isg", "0", FCVAR_CHEAT, "Is the ISG flag set?\n");
 
 // This feature enables the ISG setting and HUD features
 class ISGFeature : public Feature
@@ -52,14 +55,6 @@ CON_COMMAND_F(y_spt_set_isg,
 		Warning("y_spt_set_isg has no effect\n");
 }
 
-void ISGFeature::LoadFeature()
-{
-	if (ORIG_MiddleOfRecheck_ov_element)
-		InitCommand(y_spt_set_isg);
-}
-
-void ISGFeature::UnloadFeature() {}
-
 bool IsISGActive()
 {
 	if (spt_isg.isgFlagPtr)
@@ -67,6 +62,20 @@ bool IsISGActive()
 	else
 		return false;
 }
+
+void ISGFeature::LoadFeature()
+{
+	if (ORIG_MiddleOfRecheck_ov_element)
+	{
+		InitCommand(y_spt_set_isg);
+
+		AddHudCallback(
+		    "isg", []() { spt_hud.DrawTopHudElement(L"isg: %d", IsISGActive()); }, y_spt_hud_isg);
+	}
+}
+
+void ISGFeature::UnloadFeature() {}
+
 #else
 
 bool IsISGActive()
