--- spt/features/isg.hpp
@@ -1,3 +1,0 @@
-#pragma once
-
-bool IsISGActive();
