+++ spt/features/overlay.cpp
@@ -1,7 +1,6 @@
 #include "stdafx.h"
-#ifndef OE
-#include "overlay.hpp"
-#include "..\overlay\overlay-renderer.hpp"
+
+#include "convar.hpp"
 
 ConVar _y_spt_overlay("_y_spt_overlay",
                       "0",
@@ -34,6 +33,20 @@ ConVar _y_spt_overlay_fov("_y_spt_overlay_fov",
                           true,
                           140.0f);
 ConVar _y_spt_overlay_swap("_y_spt_overlay_swap", "0", FCVAR_CHEAT, "Swap alternate view and main screen?\n");
+ConVar _y_spt_overlay_crosshair_size("_y_spt_overlay_crosshair_size", "10", FCVAR_CHEAT, "Overlay crosshair size.");
+ConVar _y_spt_overlay_crosshair_thickness("_y_spt_overlay_crosshair_thickness",
+                                          "1",
+                                          FCVAR_CHEAT,
+                                          "Overlay crosshair thickness.");
+ConVar _y_spt_overlay_crosshair_color("_y_spt_overlay_crosshair_color",
+                                      "0 255 0 255",
+                                      FCVAR_CHEAT,
+                                      "Overlay crosshair RGBA color.");
+
+#if defined(SSDK2007)
+#include "overlay.hpp"
+#include "hud.hpp"
+#include "..\overlay\overlay-renderer.hpp"
 
 Overlay spt_overlay;
 
@@ -58,10 +71,49 @@ void Overlay::LoadFeature()
 		InitConcommandBase(_y_spt_overlay_width);
 		InitConcommandBase(_y_spt_overlay_fov);
 		InitConcommandBase(_y_spt_overlay_swap);
+
+		bool result = spt_hud.AddHudCallback(HudCallback(
+		    "overlay", std::bind(&Overlay::DrawCrosshair, this), []() { return true; }, true));
+
+		if (result)
+		{
+			InitConcommandBase(_y_spt_overlay_crosshair_size);
+			InitConcommandBase(_y_spt_overlay_crosshair_thickness);
+			InitConcommandBase(_y_spt_overlay_crosshair_color);
+		}
 	}
 }
 
 void Overlay::UnloadFeature() {}
+
+void Overlay::DrawCrosshair()
+{
+	static std::string color = "";
+	static int r = 0, g = 0, b = 0, a = 0;
+
+	if (strcmp(color.c_str(), _y_spt_overlay_crosshair_color.GetString()) != 0)
+	{
+		color = _y_spt_overlay_crosshair_color.GetString();
+		sscanf(color.c_str(), "%d %d %d %d", &r, &g, &b, &a);
+	}
+
+	vrect_t* screen = (vrect_t*)screenRect;
+	spt_hud.surface->DrawSetColor(r, g, b, a);
+	int x = screen->x + screen->width / 2;
+	int y = screen->y + screen->height / 2;
+	int width = _y_spt_overlay_crosshair_size.GetInt();
+	int thickness = _y_spt_overlay_crosshair_thickness.GetInt();
+
+	if (thickness > width)
+		std::swap(thickness, width);
+
+	spt_hud.surface->DrawFilledRect(x - thickness / 2, y - width / 2, x + thickness / 2 + 1, y + width / 2 + 1);
+	spt_hud.surface->DrawFilledRect(x - width / 2, y - thickness / 2, x - thickness / 2, y + thickness / 2 + 1);
+	spt_hud.surface->DrawFilledRect(x + thickness / 2 + 1,
+	                                y - thickness / 2,
+	                                x + width / 2 + 1,
+	                                y + thickness / 2 + 1);
+}
 
 void __fastcall Overlay::HOOKED_CViewRender__RenderView(void* thisptr,
                                                         int edx,
